Rc/3.4.+

+++ android/src/main/java/io/agora/rtc/react/RCTAgoraRtcTextureViewManager.kt
@@ -33,7 +33,7 @@ class RCTAgoraRtcTextureViewManager : SimpleViewManager<RtcTextureView>() {
   fun setData(view: RtcTextureView, data: ReadableMap) {
     data.toHashMap().let { map ->
       val channel = (map["channelId"] as? String)?.let { getChannel(it) }
-      getEngine()?.let { view.setData(it, channel, (map["uid"] as Number).toInt()) }
+      getEngine()?.let { view.setData(it, channel, map["uid"] as Number) }
     }
   }
